Training Classifier
Given a dataset that consists of 2000 cases of digits from 0 to 9 (i.e
200 numbers for each digit). We would like to treat half of the numbers
as training dataset and the other half as the test data set.
After splitting the original data set, we train our classifier with the
use of the training set. Here we are using the Principal Component
Analysis (PCA) and Linear Regression. After training the classifier
with the “m” Principal Components we are testing our algorithm on the
other (testing) dataset.
In order to minimize the number of misclassifications, we are running
the algorithm for different PCs, and observe on the graph the cases of
Underfitting, Overfitting and the optimal solution